delete order fix

+++ flaskProject/routes/api.py
@@ -117,6 +117,16 @@ def add_order_api():
 #Delete order
 @api_bp.route('/api/orders/<int:order_id>', methods=['DELETE'])
 def delete_order_api(order_id):
+    if 'user_id' not in session:
+        return jsonify({
+            "status": "error",
+            "message": "You must be logged in to access this resource."
+        }), 401
+    if session.get('is_admin') is not True:  # Перевірка, чи є користувач адміністратором
+        return jsonify({
+            "status": "error",
+            "message": "You do not have admin rights."
+        }), 403
     try:
         delete_order(order_id)
         return jsonify({'message': 'Order deleted successfully'}), 200
